Some updates

+++ basicApp/src/main/java/feyz/com/basicApp/Business/Abstracts/UserService.java
@@ -6,8 +6,8 @@ import java.util.List;
 
 public interface UserService {
 List<User> GetAll();
-User GetById(int id);
-void DeleteById(int id);
+User GetById(Long id);
+void DeleteById(Long id);
 User Add(User user);
 User Update(User user);
